requirements.txt
requests~=2.25.1
bs4~=0.0.1
beautifulsoup4~=4.9.3
mysql-connector-python
